Refactor: 불필요한 중복 탐색 로직 제거
시간 복잡도 O(n²) -> O(n) 최적화
